@@ -1,7 +1,7 @@
 name: google_maps_flutter_heatmap
 description: A Flutter plugin for integrating Google Maps in iOS and Android applications. with heatmap support
 homepage: https://github.com/bmabir17/google_maps_flutter_heatmap
-version: 0.1.1+3
+version: 0.1.2+3
 
 dependencies:
   flutter:
@@ -27,5 +27,4 @@ flutter:
         pluginClass: FLTGoogleMapsPlugin
 
 environment:
-  sdk: ">=2.0.0-dev.47.0 <3.0.0"
-  flutter: ">=1.10.0 <2.0.0"
+  sdk: '>=3.2.6 <4.0.0'
