@@ -6,9 +6,9 @@
 
 import 'package:flutter/material.dart';
 import 'package:google_maps_flutter_heatmap/google_maps_flutter_heatmap.dart';
-import 'page.dart';
+import 'package:google_maps_flutter_heatmap_example/page.dart' as p;
 
-class PaddingPage extends Page {
+class PaddingPage extends p.Page {
   PaddingPage() : super(const Icon(Icons.map), 'Add padding to the map');
 
   @override
@@ -27,7 +27,7 @@ class MarkerIconsBody extends StatefulWidget {
 const LatLng _kMapCenter = LatLng(52.4478, -3.5402);
 
 class MarkerIconsBodyState extends State<MarkerIconsBody> {
-  GoogleMapController controller;
+  late GoogleMapController controller;
 
   EdgeInsets _padding = const EdgeInsets.all(0);
 
@@ -147,19 +147,19 @@ class MarkerIconsBodyState extends State<MarkerIconsBody> {
       child: Row(
         mainAxisAlignment: MainAxisAlignment.spaceEvenly,
         children: <Widget>[
-          FlatButton(
+          FilledButton(
             child: const Text("Set Padding"),
             onPressed: () {
               setState(() {
                 _padding = EdgeInsets.fromLTRB(
-                    double.tryParse(_leftController.value?.text) ?? 0,
-                    double.tryParse(_topController.value?.text) ?? 0,
-                    double.tryParse(_rightController.value?.text) ?? 0,
-                    double.tryParse(_bottomController.value?.text) ?? 0);
+                    double.tryParse(_leftController.value.text) ?? 0,
+                    double.tryParse(_topController.value.text) ?? 0,
+                    double.tryParse(_rightController.value.text) ?? 0,
+                    double.tryParse(_bottomController.value.text) ?? 0);
               });
             },
           ),
-          FlatButton(
+          FilledButton(
             child: const Text("Reset Padding"),
             onPressed: () {
               setState(() {
