@@ -7,9 +7,9 @@
 import 'package:flutter/material.dart';
 import 'package:google_maps_flutter_heatmap/google_maps_flutter_heatmap.dart';
 
-import 'page.dart';
+import 'package:google_maps_flutter_heatmap_example/page.dart' as p;
 
-class PlaceCirclePage extends Page {
+class PlaceCirclePage extends p.Page {
   PlaceCirclePage() : super(const Icon(Icons.linear_scale), 'Place circle');
 
   @override
@@ -28,10 +28,10 @@ class PlaceCircleBody extends StatefulWidget {
 class PlaceCircleBodyState extends State<PlaceCircleBody> {
   PlaceCircleBodyState();
 
-  GoogleMapController controller;
+  late GoogleMapController controller;
   Map<CircleId, Circle> circles = <CircleId, Circle>{};
   int _circleIdCounter = 1;
-  CircleId selectedCircle;
+  late CircleId? selectedCircle;
 
   // Values when toggling circle color
   int fillColorsIndex = 0;
@@ -101,36 +101,36 @@ class PlaceCircleBodyState extends State<PlaceCircleBody> {
   }
 
   void _toggleVisible() {
-    final Circle circle = circles[selectedCircle];
+    final Circle circle = circles[selectedCircle]!;
     setState(() {
-      circles[selectedCircle] = circle.copyWith(
+      circles[selectedCircle!] = circle.copyWith(
         visibleParam: !circle.visible,
       );
     });
   }
 
   void _changeFillColor() {
-    final Circle circle = circles[selectedCircle];
+    final Circle circle = circles[selectedCircle]!;
     setState(() {
-      circles[selectedCircle] = circle.copyWith(
+      circles[selectedCircle!] = circle.copyWith(
         fillColorParam: colors[++fillColorsIndex % colors.length],
       );
     });
   }
 
   void _changeStrokeColor() {
-    final Circle circle = circles[selectedCircle];
+    final Circle circle = circles[selectedCircle]!;
     setState(() {
-      circles[selectedCircle] = circle.copyWith(
+      circles[selectedCircle!] = circle.copyWith(
         strokeColorParam: colors[++strokeColorsIndex % colors.length],
       );
     });
   }
 
   void _changeStrokeWidth() {
-    final Circle circle = circles[selectedCircle];
+    final Circle circle = circles[selectedCircle]!;
     setState(() {
-      circles[selectedCircle] = circle.copyWith(
+      circles[selectedCircle!] = circle.copyWith(
         strokeWidthParam: widths[++widthsIndex % widths.length],
       );
     });
@@ -165,15 +165,15 @@ class PlaceCircleBodyState extends State<PlaceCircleBody> {
                   children: <Widget>[
                     Column(
                       children: <Widget>[
-                        FlatButton(
+                        FilledButton(
                           child: const Text('add'),
                           onPressed: _add,
                         ),
-                        FlatButton(
+                        FilledButton(
                           child: const Text('remove'),
                           onPressed: (selectedCircle == null) ? null : _remove,
                         ),
-                        FlatButton(
+                        FilledButton(
                           child: const Text('toggle visible'),
                           onPressed:
                               (selectedCircle == null) ? null : _toggleVisible,
@@ -182,19 +182,19 @@ class PlaceCircleBodyState extends State<PlaceCircleBody> {
                     ),
                     Column(
                       children: <Widget>[
-                        FlatButton(
+                        FilledButton(
                           child: const Text('change stroke width'),
                           onPressed: (selectedCircle == null)
                               ? null
                               : _changeStrokeWidth,
                         ),
-                        FlatButton(
+                        FilledButton(
                           child: const Text('change stroke color'),
                           onPressed: (selectedCircle == null)
                               ? null
                               : _changeStrokeColor,
                         ),
-                        FlatButton(
+                        FilledButton(
                           child: const Text('change fill color'),
                           onPressed: (selectedCircle == null)
                               ? null
