@@ -13,13 +13,13 @@ class CircleId {
   CircleId(this.value) : assert(value != null);
 
   /// value of the [CircleId].
-  final String value;
+  final String? value;
 
   @override
   bool operator ==(Object other) {
     if (identical(this, other)) return true;
     if (other.runtimeType != runtimeType) return false;
-    final CircleId typedOther = other;
+    final CircleId typedOther = other as CircleId;
     return value == typedOther.value;
   }
 
@@ -37,7 +37,7 @@ class CircleId {
 class Circle {
   /// Creates an immutable representation of a [Circle] to draw on [GoogleMap].
   const Circle({
-    @required this.circleId,
+    required this.circleId,
     this.consumeTapEvents = false,
     this.fillColor = Colors.transparent,
     this.center = const LatLng(0.0, 0.0),
@@ -87,20 +87,20 @@ class Circle {
   final int zIndex;
 
   /// Callbacks to receive tap events for circle placed on this map.
-  final VoidCallback onTap;
+  final VoidCallback? onTap;
 
   /// Creates a new [Circle] object whose values are the same as this instance,
   /// unless overwritten by the specified parameters.
   Circle copyWith({
-    bool consumeTapEventsParam,
-    Color fillColorParam,
-    LatLng centerParam,
-    double radiusParam,
-    Color strokeColorParam,
-    int strokeWidthParam,
-    bool visibleParam,
-    int zIndexParam,
-    VoidCallback onTapParam,
+    bool? consumeTapEventsParam,
+    Color? fillColorParam,
+    LatLng? centerParam,
+    double? radiusParam,
+    Color? strokeColorParam,
+    int? strokeWidthParam,
+    bool? visibleParam,
+    int? zIndexParam,
+    VoidCallback? onTapParam,
   }) {
     return Circle(
       circleId: circleId,
@@ -145,7 +145,7 @@ class Circle {
   bool operator ==(Object other) {
     if (identical(this, other)) return true;
     if (other.runtimeType != runtimeType) return false;
-    final Circle typedOther = other;
+    final Circle typedOther = other as Circle;
     return circleId == typedOther.circleId &&
         consumeTapEvents == typedOther.consumeTapEvents &&
         fillColor == typedOther.fillColor &&
@@ -162,7 +162,7 @@ class Circle {
   int get hashCode => circleId.hashCode;
 }
 
-Map<CircleId, Circle> _keyByCircleId(Iterable<Circle> circles) {
+Map<CircleId, Circle> _keyByCircleId(Iterable<Circle>? circles) {
   if (circles == null) {
     return <CircleId, Circle>{};
   }
@@ -170,9 +170,9 @@ Map<CircleId, Circle> _keyByCircleId(Iterable<Circle> circles) {
       MapEntry<CircleId, Circle>(circle.circleId, circle.clone())));
 }
 
-List<Map<String, dynamic>> _serializeCircleSet(Set<Circle> circles) {
+List<Map<String, dynamic>>? _serializeCircleSet(Set<Circle?>? circles) {
   if (circles == null) {
     return null;
   }
-  return circles.map<Map<String, dynamic>>((Circle p) => p._toJson()).toList();
+  return circles.map<Map<String, dynamic>>((Circle? p) => p!._toJson()).toList();
 }
