@@ -14,8 +14,7 @@ class LatLng {
   /// The longitude is normalized to the half-open interval from -180.0
   /// (inclusive) to +180.0 (exclusive)
   const LatLng(double latitude, double longitude)
-      : assert(latitude != null),
-        assert(longitude != null),
+      : 
         latitude =
             (latitude < -90.0 ? -90.0 : (90.0 < latitude ? 90.0 : latitude)),
         longitude = (longitude + 180.0) % 360.0 - 180.0;
@@ -30,7 +29,7 @@ class LatLng {
     return <double>[latitude, longitude];
   }
 
-  static LatLng _fromJson(dynamic json) {
+  static LatLng? _fromJson(dynamic json) {
     if (json == null) {
       return null;
     }
@@ -46,7 +45,7 @@ class LatLng {
   }
 
   @override
-  int get hashCode => hashValues(latitude, longitude);
+  int get hashCode => Object.hash(latitude, longitude);
 }
 
 /// A latitude/longitude aligned rectangle.
@@ -62,19 +61,19 @@ class LatLngBounds {
   ///
   /// The latitude of the southwest corner cannot be larger than the
   /// latitude of the northeast corner.
-  LatLngBounds({@required this.southwest, @required this.northeast})
+  LatLngBounds({required this.southwest, required this.northeast})
       : assert(southwest != null),
         assert(northeast != null),
-        assert(southwest.latitude <= northeast.latitude);
+        assert(southwest!.latitude <= northeast!.latitude);
 
   /// The southwest corner of the rectangle.
-  final LatLng southwest;
+  final LatLng? southwest;
 
   /// The northeast corner of the rectangle.
-  final LatLng northeast;
+  final LatLng? northeast;
 
   dynamic _toList() {
-    return <dynamic>[southwest._toJson(), northeast._toJson()];
+    return <dynamic>[southwest!._toJson(), northeast!._toJson()];
   }
 
   /// Returns whether this rectangle contains the given [LatLng].
@@ -84,20 +83,20 @@ class LatLngBounds {
   }
 
   bool _containsLatitude(double lat) {
-    return (southwest.latitude <= lat) && (lat <= northeast.latitude);
+    return (southwest!.latitude <= lat) && (lat <= northeast!.latitude);
   }
 
   bool _containsLongitude(double lng) {
-    if (southwest.longitude <= northeast.longitude) {
-      return southwest.longitude <= lng && lng <= northeast.longitude;
+    if (southwest!.longitude <= northeast!.longitude) {
+      return southwest!.longitude <= lng && lng <= northeast!.longitude;
     } else {
-      return southwest.longitude <= lng || lng <= northeast.longitude;
+      return southwest!.longitude <= lng || lng <= northeast!.longitude;
     }
   }
 
   /// Converts a list to [LatLngBounds].
   @visibleForTesting
-  static LatLngBounds fromList(dynamic json) {
+  static LatLngBounds? fromList(dynamic json) {
     if (json == null) {
       return null;
     }
