@@ -4,7 +4,7 @@
 
 part of google_maps_flutter_heatmap;
 
-dynamic _offsetToJson(Offset offset) {
+dynamic _offsetToJson(Offset? offset) {
   if (offset == null) {
     return null;
   }
@@ -27,12 +27,12 @@ class InfoWindow {
   /// Text displayed in an info window when the user taps the marker.
   ///
   /// A null value means no title.
-  final String title;
+  final String? title;
 
   /// Additional text displayed below the [title].
   ///
   /// A null value means no additional text.
-  final String snippet;
+  final String? snippet;
 
   /// The icon image point that will be the anchor of the info window when
   /// displayed.
@@ -40,18 +40,18 @@ class InfoWindow {
   /// The image point is specified in normalized coordinates: An anchor of
   /// (0.0, 0.0) means the top left corner of the image. An anchor
   /// of (1.0, 1.0) means the bottom right corner of the image.
-  final Offset anchor;
+  final Offset? anchor;
 
   /// onTap callback for this [InfoWindow].
-  final VoidCallback onTap;
+  final VoidCallback? onTap;
 
   /// Creates a new [InfoWindow] object whose values are the same as this instance,
   /// unless overwritten by the specified parameters.
   InfoWindow copyWith({
-    String titleParam,
-    String snippetParam,
-    Offset anchorParam,
-    VoidCallback onTapParam,
+    String? titleParam,
+    String? snippetParam,
+    Offset? anchorParam,
+    VoidCallback? onTapParam,
   }) {
     return InfoWindow(
       title: titleParam ?? title,
@@ -81,7 +81,7 @@ class InfoWindow {
   bool operator ==(Object other) {
     if (identical(this, other)) return true;
     if (other.runtimeType != runtimeType) return false;
-    final InfoWindow typedOther = other;
+    final InfoWindow typedOther = other as InfoWindow;
     return title == typedOther.title &&
         snippet == typedOther.snippet &&
         anchor == typedOther.anchor;
@@ -105,13 +105,13 @@ class MarkerId {
   MarkerId(this.value) : assert(value != null);
 
   /// value of the [MarkerId].
-  final String value;
+  final String? value;
 
   @override
   bool operator ==(Object other) {
     if (identical(this, other)) return true;
     if (other.runtimeType != runtimeType) return false;
-    final MarkerId typedOther = other;
+    final MarkerId typedOther = other as MarkerId;
     return value == typedOther.value;
   }
 
@@ -151,7 +151,7 @@ class Marker {
   /// * reports [onTap] events
   /// * reports [onDragEnd] events
   const Marker({
-    @required this.markerId,
+    required this.markerId,
     this.alpha = 1.0,
     this.anchor = const Offset(0.5, 1.0),
     this.consumeTapEvents = false,
@@ -173,7 +173,7 @@ class Marker {
   /// The opacity of the marker, between 0.0 and 1.0 inclusive.
   ///
   /// 0.0 means fully transparent, 1.0 means fully opaque.
-  final double alpha;
+  final double? alpha;
 
   /// The icon image point that will be placed at the [position] of the marker.
   ///
@@ -219,27 +219,27 @@ class Marker {
   final double zIndex;
 
   /// Callbacks to receive tap events for markers placed on this map.
-  final VoidCallback onTap;
+  final VoidCallback? onTap;
 
   /// Signature reporting the new [LatLng] at the end of a drag event.
-  final ValueChanged<LatLng> onDragEnd;
+  final ValueChanged<LatLng>? onDragEnd;
 
   /// Creates a new [Marker] object whose values are the same as this instance,
   /// unless overwritten by the specified parameters.
   Marker copyWith({
-    double alphaParam,
-    Offset anchorParam,
-    bool consumeTapEventsParam,
-    bool draggableParam,
-    bool flatParam,
-    BitmapDescriptor iconParam,
-    InfoWindow infoWindowParam,
-    LatLng positionParam,
-    double rotationParam,
-    bool visibleParam,
-    double zIndexParam,
-    VoidCallback onTapParam,
-    ValueChanged<LatLng> onDragEndParam,
+    double? alphaParam,
+    Offset? anchorParam,
+    bool? consumeTapEventsParam,
+    bool? draggableParam,
+    bool? flatParam,
+    BitmapDescriptor? iconParam,
+    InfoWindow? infoWindowParam,
+    LatLng? positionParam,
+    double? rotationParam,
+    bool? visibleParam,
+    double? zIndexParam,
+    VoidCallback? onTapParam,
+    ValueChanged<LatLng>? onDragEndParam,
   }) {
     return Marker(
       markerId: markerId,
@@ -277,9 +277,9 @@ class Marker {
     addIfPresent('consumeTapEvents', consumeTapEvents);
     addIfPresent('draggable', draggable);
     addIfPresent('flat', flat);
-    addIfPresent('icon', icon?._toJson());
-    addIfPresent('infoWindow', infoWindow?._toJson());
-    addIfPresent('position', position?._toJson());
+    addIfPresent('icon', icon._toJson());
+    addIfPresent('infoWindow', infoWindow._toJson());
+    addIfPresent('position', position._toJson());
     addIfPresent('rotation', rotation);
     addIfPresent('visible', visible);
     addIfPresent('zIndex', zIndex);
@@ -290,7 +290,7 @@ class Marker {
   bool operator ==(Object other) {
     if (identical(this, other)) return true;
     if (other.runtimeType != runtimeType) return false;
-    final Marker typedOther = other;
+    final Marker typedOther = other as Marker;
     return markerId == typedOther.markerId &&
         alpha == typedOther.alpha &&
         anchor == typedOther.anchor &&
@@ -318,7 +318,7 @@ class Marker {
   }
 }
 
-Map<MarkerId, Marker> _keyByMarkerId(Iterable<Marker> markers) {
+Map<MarkerId, Marker> _keyByMarkerId(Iterable<Marker>? markers) {
   if (markers == null) {
     return <MarkerId, Marker>{};
   }
@@ -326,9 +326,9 @@ Map<MarkerId, Marker> _keyByMarkerId(Iterable<Marker> markers) {
       MapEntry<MarkerId, Marker>(marker.markerId, marker.clone())));
 }
 
-List<Map<String, dynamic>> _serializeMarkerSet(Set<Marker> markers) {
+List<Map<String, dynamic>>? _serializeMarkerSet(Set<Marker?>? markers) {
   if (markers == null) {
     return null;
   }
-  return markers.map<Map<String, dynamic>>((Marker m) => m._toJson()).toList();
+  return markers.map<Map<String, dynamic>>((Marker? m) => m!._toJson()).toList();
 }
