@@ -13,13 +13,13 @@ class PolygonId {
   PolygonId(this.value) : assert(value != null);
 
   /// value of the [PolygonId].
-  final String value;
+  final String? value;
 
   @override
   bool operator ==(Object other) {
     if (identical(this, other)) return true;
     if (other.runtimeType != runtimeType) return false;
-    final PolygonId typedOther = other;
+    final PolygonId typedOther = other as PolygonId;
     return value == typedOther.value;
   }
 
@@ -37,7 +37,7 @@ class PolygonId {
 class Polygon {
   /// Creates an immutable representation of a polygon through geographical locations on the map.
   const Polygon({
-    @required this.polygonId,
+    required this.polygonId,
     this.consumeTapEvents = false,
     this.fillColor = Colors.black,
     this.geodesic = false,
@@ -71,7 +71,7 @@ class Polygon {
   ///
   /// Line segments are drawn between consecutive points. A polygon is not closed by
   /// default; to form a closed polygon, the start and end points must be the same.
-  final List<LatLng> points;
+  final List<LatLng>? points;
 
   /// True if the marker is visible.
   final bool visible;
@@ -93,20 +93,20 @@ class Polygon {
   final int zIndex;
 
   /// Callbacks to receive tap events for polygon placed on this map.
-  final VoidCallback onTap;
+  final VoidCallback? onTap;
 
   /// Creates a new [Polygon] object whose values are the same as this instance,
   /// unless overwritten by the specified parameters.
   Polygon copyWith({
-    bool consumeTapEventsParam,
-    Color fillColorParam,
-    bool geodesicParam,
-    List<LatLng> pointsParam,
-    Color strokeColorParam,
-    int strokeWidthParam,
-    bool visibleParam,
-    int zIndexParam,
-    VoidCallback onTapParam,
+    bool? consumeTapEventsParam,
+    Color? fillColorParam,
+    bool? geodesicParam,
+    List<LatLng>? pointsParam,
+    Color? strokeColorParam,
+    int? strokeWidthParam,
+    bool? visibleParam,
+    int? zIndexParam,
+    VoidCallback? onTapParam,
   }) {
     return Polygon(
       polygonId: polygonId,
@@ -124,7 +124,10 @@ class Polygon {
 
   /// Creates a new [Polygon] object whose values are the same as this instance.
   Polygon clone() {
-    return copyWith(pointsParam: List<LatLng>.of(points));
+    if (points == null) {
+      return copyWith();
+    }
+    return copyWith(pointsParam: List<LatLng>.of(points!));
   }
 
   dynamic _toJson() {
@@ -156,7 +159,7 @@ class Polygon {
   bool operator ==(Object other) {
     if (identical(this, other)) return true;
     if (other.runtimeType != runtimeType) return false;
-    final Polygon typedOther = other;
+    final Polygon typedOther = other as Polygon;
     return polygonId == typedOther.polygonId &&
         consumeTapEvents == typedOther.consumeTapEvents &&
         fillColor == typedOther.fillColor &&
@@ -174,14 +177,14 @@ class Polygon {
 
   dynamic _pointsToJson() {
     final List<dynamic> result = <dynamic>[];
-    for (final LatLng point in points) {
+    for (final LatLng point in points!) {
       result.add(point._toJson());
     }
     return result;
   }
 }
 
-Map<PolygonId, Polygon> _keyByPolygonId(Iterable<Polygon> polygons) {
+Map<PolygonId, Polygon> _keyByPolygonId(Iterable<Polygon>? polygons) {
   if (polygons == null) {
     return <PolygonId, Polygon>{};
   }
@@ -189,11 +192,11 @@ Map<PolygonId, Polygon> _keyByPolygonId(Iterable<Polygon> polygons) {
       MapEntry<PolygonId, Polygon>(polygon.polygonId, polygon.clone())));
 }
 
-List<Map<String, dynamic>> _serializePolygonSet(Set<Polygon> polygons) {
+List<Map<String, dynamic>>? _serializePolygonSet(Set<Polygon?>? polygons) {
   if (polygons == null) {
     return null;
   }
   return polygons
-      .map<Map<String, dynamic>>((Polygon p) => p._toJson())
+      .map<Map<String, dynamic>>((Polygon? p) => p!._toJson())
       .toList();
 }
