@@ -9,7 +9,7 @@ part of google_maps_flutter_heatmap;
 /// Used in [GoogleMapController] when the map is updated.
 class _PolylineUpdates {
   /// Computes [_PolylineUpdates] given previous and current [Polyline]s.
-  _PolylineUpdates.from(Set<Polyline> previous, Set<Polyline> current) {
+  _PolylineUpdates.from(Set<Polyline>? previous, Set<Polyline>? current) {
     if (previous == null) {
       previous = Set<Polyline>.identity();
     }
@@ -26,26 +26,29 @@ class _PolylineUpdates {
     final Set<PolylineId> prevPolylineIds = previousPolylines.keys.toSet();
     final Set<PolylineId> currentPolylineIds = currentPolylines.keys.toSet();
 
-    Polyline idToCurrentPolyline(PolylineId id) {
+    Polyline? idToCurrentPolyline(PolylineId id) {
       return currentPolylines[id];
     }
 
     final Set<PolylineId> _polylineIdsToRemove =
         prevPolylineIds.difference(currentPolylineIds);
 
-    final Set<Polyline> _polylinesToAdd = currentPolylineIds
+    final Set<Polyline?> _polylinesToAdd = currentPolylineIds
         .difference(prevPolylineIds)
         .map(idToCurrentPolyline)
         .toSet();
 
     /// Returns `true` if [current] is not equals to previous one with the
     /// same id.
-    bool hasChanged(Polyline current) {
-      final Polyline previous = previousPolylines[current.polylineId];
+    bool hasChanged(Polyline? current) {
+      if (current == null) {
+        return false;
+      }
+      final Polyline? previous = previousPolylines[current.polylineId];
       return current != previous;
     }
 
-    final Set<Polyline> _polylinesToChange = currentPolylineIds
+    final Set<Polyline?> _polylinesToChange = currentPolylineIds
         .intersection(prevPolylineIds)
         .map(idToCurrentPolyline)
         .where(hasChanged)
@@ -56,9 +59,9 @@ class _PolylineUpdates {
     polylinesToChange = _polylinesToChange;
   }
 
-  Set<Polyline> polylinesToAdd;
-  Set<PolylineId> polylineIdsToRemove;
-  Set<Polyline> polylinesToChange;
+  late Set<Polyline?>? polylinesToAdd;
+  late Set<PolylineId> polylineIdsToRemove;
+  late Set<Polyline?> polylinesToChange;
 
   Map<String, dynamic> _toMap() {
     final Map<String, dynamic> updateMap = <String, dynamic>{};
@@ -81,7 +84,7 @@ class _PolylineUpdates {
   bool operator ==(Object other) {
     if (identical(this, other)) return true;
     if (other.runtimeType != runtimeType) return false;
-    final _PolylineUpdates typedOther = other;
+    final _PolylineUpdates typedOther = other as _PolylineUpdates;
     return setEquals(polylinesToAdd, typedOther.polylinesToAdd) &&
         setEquals(polylineIdsToRemove, typedOther.polylineIdsToRemove) &&
         setEquals(polylinesToChange, typedOther.polylinesToChange);
@@ -89,7 +92,7 @@ class _PolylineUpdates {
 
   @override
   int get hashCode =>
-      hashValues(polylinesToAdd, polylineIdsToRemove, polylinesToChange);
+      Object.hash(polylinesToAdd, polylineIdsToRemove, polylinesToChange);
 
   @override
   String toString() {
